# video-message-app/D_ID_API_SPEC.yaml
openapi: 3.0.3
info:
  title: D-ID Video Generation API
  description: |
    Backend API wrapper for D-ID video generation service.

    ## Features
    - Upload images and audio files to D-ID
    - Generate talking avatar videos with lip-sync
    - Monitor video generation status
    - Health check endpoint

    ## Authentication
    D-ID API key is configured server-side in `.env` file.
    No authentication required for these wrapper endpoints.

    ## Rate Limits
    - Free Plan: 10 requests/minute
    - Lite Plan: 30 requests/minute
    - Pro Plan: 60 requests/minute

  version: 1.0.0
  contact:
    name: Video Message App Support
    email: support@video-message-app.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:55433/api/d-id
    description: Local development
  - url: http://3.115.141.166:55433/api/d-id
    description: EC2 production

tags:
  - name: Upload
    description: Upload image and audio files to D-ID
  - name: Video
    description: Generate and manage videos
  - name: Health
    description: Service health check

paths:
  /health:
    get:
      tags:
        - Health
      summary: Health check
      description: Check if D-ID service is configured and healthy
      operationId: healthCheck
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
              example:
                status: healthy
                service: d-id
                api_key_configured: true

  /upload-source-image:
    post:
      tags:
        - Upload
      summary: Upload source image
      description: |
        Upload a portrait image to D-ID for video generation.

        **Requirements**:
        - Format: JPEG, PNG, WebP
        - Max size: 10 MB
        - Recommended resolution: 512x512 to 1024x1024
        - Portrait photo with clear face
      operationId: uploadSourceImage
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: Image file to upload
      responses:
        '200':
          description: Image uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
              example:
                url: https://d-id-prod-us.s3.amazonaws.com/images/img_abc123xyz.jpg
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /upload-audio:
    post:
      tags:
        - Upload
      summary: Upload audio file
      description: |
        Upload an audio file to D-ID for video generation.

        **Requirements**:
        - Format: WAV, MP3, MP4, FLAC, M4A
        - Max size: 30 MB
        - Max duration: 5 minutes
        - Recommended: 16kHz mono WAV
      operationId: uploadAudio
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: Audio file to upload
      responses:
        '200':
          description: Audio uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
              example:
                url: https://d-id-prod-us.s3.amazonaws.com/audios/aud_abc123xyz.wav
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /generate-video:
    post:
      tags:
        - Video
      summary: Generate talking avatar video
      description: |
        Generate a lip-sync video from uploaded image and audio.

        **Process**:
        1. Upload image and audio using respective endpoints
        2. Call this endpoint with the returned URLs
        3. Poll status endpoint until video is ready (30-60 seconds)

        **Credits**:
        - Audio-to-video (lip-sync): 1 credit per video
      operationId: generateVideo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VideoGenerationRequest'
            example:
              audio_url: https://d-id-prod-us.s3.amazonaws.com/audios/aud_abc123xyz.wav
              source_url: https://d-id-prod-us.s3.amazonaws.com/images/img_abc123xyz.jpg
      responses:
        '200':
          description: Video generation started successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoGenerationResponse'
              example:
                id: tlk_abc123xyz
                status: done
                result_url: https://d-id-talks-prod.s3.amazonaws.com/tlk_abc123xyz/video.mp4
                created_at: '2025-11-07T12:34:56.789Z'
        '400':
          $ref: '#/components/responses/BadRequest'
        '402':
          $ref: '#/components/responses/PaymentRequired'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /talk-status/{talk_id}:
    get:
      tags:
        - Video
      summary: Get video generation status
      description: |
        Poll this endpoint to check video generation progress.

        **Polling Strategy**:
        - Poll every 5 seconds
        - Max attempts: 60 (5 minutes)
        - Status values: `created`, `processing`, `done`, `error`, `rejected`
      operationId: getTalkStatus
      parameters:
        - name: talk_id
          in: path
          required: true
          description: D-ID talk ID returned from generate-video endpoint
          schema:
            type: string
            pattern: '^tlk_[a-zA-Z0-9]+$'
          example: tlk_abc123xyz
      responses:
        '200':
          description: Status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TalkStatusResponse'
              examples:
                processing:
                  summary: Video is being generated
                  value:
                    id: tlk_abc123xyz
                    status: processing
                    created_at: '2025-11-07T12:34:56.789Z'
                done:
                  summary: Video is ready
                  value:
                    id: tlk_abc123xyz
                    status: done
                    result_url: https://d-id-talks-prod.s3.amazonaws.com/tlk_abc123xyz/video.mp4
                    created_at: '2025-11-07T12:34:56.789Z'
                    duration: 12.5
                error:
                  summary: Video generation failed
                  value:
                    id: tlk_abc123xyz
                    status: error
                    error:
                      kind: InvalidRequest
                      description: Audio URL is not accessible
                    created_at: '2025-11-07T12:34:56.789Z'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  schemas:
    HealthResponse:
      type: object
      required:
        - status
        - service
        - api_key_configured
      properties:
        status:
          type: string
          enum: [healthy, unhealthy]
          description: Overall service health status
        service:
          type: string
          description: Service name
          example: d-id
        api_key_configured:
          type: boolean
          description: Whether D-ID API key is configured

    UploadResponse:
      type: object
      required:
        - url
      properties:
        url:
          type: string
          format: uri
          description: URL of uploaded file on D-ID CDN
          example: https://d-id-prod-us.s3.amazonaws.com/images/img_abc123xyz.jpg

    VideoGenerationRequest:
      type: object
      required:
        - audio_url
        - source_url
      properties:
        audio_url:
          type: string
          format: uri
          description: URL of audio file (from upload-audio endpoint)
          example: https://d-id-prod-us.s3.amazonaws.com/audios/aud_abc123xyz.wav
        source_url:
          type: string
          format: uri
          description: URL of source image (from upload-source-image endpoint)
          example: https://d-id-prod-us.s3.amazonaws.com/images/img_abc123xyz.jpg

    VideoGenerationResponse:
      type: object
      required:
        - id
        - status
        - created_at
      properties:
        id:
          type: string
          pattern: '^tlk_[a-zA-Z0-9]+$'
          description: Unique talk ID for this video
          example: tlk_abc123xyz
        status:
          type: string
          enum: [created, processing, done, error, rejected]
          description: Current video generation status
        result_url:
          type: string
          format: uri
          description: URL of generated video (only present when status is 'done')
          example: https://d-id-talks-prod.s3.amazonaws.com/tlk_abc123xyz/video.mp4
        created_at:
          type: string
          format: date-time
          description: Timestamp when video generation was initiated
          example: '2025-11-07T12:34:56.789Z'

    TalkStatusResponse:
      type: object
      required:
        - id
        - status
        - created_at
      properties:
        id:
          type: string
          pattern: '^tlk_[a-zA-Z0-9]+$'
          description: Unique talk ID
          example: tlk_abc123xyz
        status:
          type: string
          enum: [created, processing, done, error, rejected]
          description: Current video generation status
        result_url:
          type: string
          format: uri
          description: URL of generated video (only present when status is 'done')
          example: https://d-id-talks-prod.s3.amazonaws.com/tlk_abc123xyz/video.mp4
        created_at:
          type: string
          format: date-time
          description: Timestamp when video generation was initiated
        duration:
          type: number
          format: float
          description: Video duration in seconds (only present when status is 'done')
          example: 12.5
        error:
          $ref: '#/components/schemas/ErrorDetails'

    ErrorDetails:
      type: object
      properties:
        kind:
          type: string
          description: Error type
          example: InvalidRequest
        description:
          type: string
          description: Human-readable error description
          example: Audio URL is not accessible
        details:
          type: object
          additionalProperties: true
          description: Additional error details

    ErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Error message
          example: 画像のアップロードに失敗しました

  responses:
    BadRequest:
      description: Invalid request (validation error)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: 画像ファイルをアップロードしてください

    PaymentRequired:
      description: Credit quota exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: Credit quota exceeded. Please upgrade your plan.

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: Talk ID not found

    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: 動画生成に失敗しました

  examples:
    SuccessfulVideoGeneration:
      summary: Successful video generation workflow
      description: |
        Complete workflow for generating a talking avatar video:

        1. Upload image:
           POST /upload-source-image
           → Returns image_url

        2. Upload audio:
           POST /upload-audio
           → Returns audio_url

        3. Generate video:
           POST /generate-video
           → Returns talk_id

        4. Poll status:
           GET /talk-status/{talk_id}
           → Returns status (processing → done)

        5. Download video:
           Video URL is available in result_url